refactor: evm reader module structure (WIP)
Also fix claim status checker
