+++ internal/evmreader/retrypolicy/retrypolicy_ethclient_delegator.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/cartesi/rollups-node/internal/evmreader"
-	"github.com/cartesi/rollups-node/internal/retry"
+	"github.com/cartesi/rollups-node/internal/services/retry"
 	"github.com/ethereum/go-ethereum/core/types"
 )
